# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Glory Kamples API",
    "description": "API Kamples — Rust + Axum + OpenAPI",
    "contact": {
      "name": "1ndoryu"
    },
    "license": {
      "name": "MIT"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/pagos/checkout": {
      "post": {
        "tags": [
          "payments"
        ],
        "operationId": "create_subscription_checkout",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSubscriptionCheckoutRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "URL de Stripe Checkout para suscripción",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaymentRedirectResponse"
                }
              }
            }
          },
          "400": {
            "description": "Plan o periodo inválido",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "No autenticado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Stripe no configurado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
    "/api/pagos/checkout-sample": {
      "post": {
        "tags": [
          "payments"
        ],
        "operationId": "create_sample_checkout",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSampleCheckoutRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "URL de Stripe Checkout para compra individual",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaymentRedirectResponse"
                }
              }
            }
          },
          "400": {
            "description": "Sample inválido o no comprable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "No autenticado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Sample no encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Stripe no configurado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
...
    "/api/pagos/portal": {
      "post": {
        "tags": [
          "payments"
        ],
        "operationId": "create_billing_portal",
        "responses": {
          "200": {
            "description": "URL del Customer Portal",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaymentRedirectResponse"
                }
              }
            }
          },
          "401": {
            "description": "No autenticado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Stripe no configurado o customer ausente",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
    "/api/pagos/webhook": {
      "post": {
        "tags": [
          "payments"
        ],
        "operationId": "payment_webhook",
        "parameters": [
          {
            "name": "stripe-signature",
            "in": "header",
            "description": "Firma HMAC Stripe",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "description": "",
          "content": {
            "application/octet-stream": {
              "schema": {
                "type": "string",
                "format": "binary"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Webhook recibido",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaymentWebhookResponse"
                }
              }
            }
          },
          "400": {
            "description": "Firma o payload inválido",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Stripe no configurado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateSampleCheckoutRequest": {
        "type": "object",
        "required": [
          "sampleId"
        ],
        "properties": {
          "sampleId": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
...
      "CreateSubscriptionCheckoutRequest": {
        "type": "object",
        "required": [
          "plan"
        ],
        "properties": {
          "periodo": {
            "allOf": [
              {
                "$ref": "#/components/schemas/PaymentPlanPeriod"
              }
            ],
            "nullable": true
          },
          "plan": {
            "$ref": "#/components/schemas/KamplesPlanId"
          }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "description": "Estructura de respuesta de error expuesta en la API",
        "required": [
          "error",
          "message"
        ],
        "properties": {
          "error": {
            "type": "string",
            "description": "Tipo de error (`not_found`, `unauthorized`, etc.)"
          },
          "message": {
            "type": "string",
            "description": "Mensaje legible para el usuario"
          }
        }
      },
...
      "KamplesPlanId": {
        "type": "string",
        "enum": [
          "free",
          "pro",
          "premium"
        ]
      },
...
      "PaymentPlanPeriod": {
        "type": "string",
        "enum": [
          "mensual",
          "anual"
        ]
      },
...
      "PaymentRedirectResponse": {
        "type": "object",
        "required": [
          "ok",
          "url"
        ],
        "properties": {
          "ok": {
            "type": "boolean"
          },
          "url": {
            "type": "string"
          }
        }
      },
      "PaymentWebhookResponse": {
        "type": "object",
        "required": [
          "recibido",
          "procesado"
        ],
        "properties": {
          "procesado": {
            "type": "boolean"
          },
          "recibido": {
            "type": "boolean"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearer_auth": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}